requirements.txt
numpy
opencv-python
easyocr==1.4.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
pandas==1.3.4
pandocfilters==1.5.0
Pillow==8.2.0
requests==2.26.0
scikit-image==0.18.3
seaborn==0.11.2
streamlit==1.0.0
imutils==0.5.4
tensorflow==2.9.2
pickle5
scikit-learn
